Completed image upload

+++ backend/src/modules/product/adapters/inbound/http/dtos/create_upload_request.rs
@@ -4,5 +4,5 @@ use serde::Deserialize;
 pub struct CreateUploadRequest {
     pub content_type: String,
     pub file_size: i64,
-    pub display_order: u8, //TODO: should I make it optional?
+    pub display_order: i32, //TODO: should I make it optional?
 }
